api/requirements.txt
fastapi[standard]
uvicorn[standard]
gunicorn
pydantic
pandas
numpy
pyarrow==18.1.0
snowflake-connector-python==3.14.0